Skip pyzstd with Python 3.14+
3.14 has builtin zstd support, so use the builtin one there.
