+++ pyproject.toml
@@ -10,7 +10,7 @@ dependencies = [
     "tabulate>=0.9.0,<0.10",
     "requests>=2.28.2,<3",
     "pydantic>=2.0,<3",
-    "pyzstd>=0.18.0,<0.19",
+    "pyzstd>=0.18.0,<0.19; python_version < '3.14'",
 ]
 
 [project.optional-dependencies]
